+++ HidingBar.lua
@@ -6,9 +6,8 @@ hidingBar.cover:Hide()
 hidingBar.cover:SetAllPoints()
 hidingBar.cover:EnableMouse(true)
 hidingBar.cover:SetFrameLevel(hidingBar:GetFrameLevel() + 10)
-hidingBar.drag = CreateFrame("FRAME", nil, UIParent)
+hidingBar.drag = CreateFrame("BUTTON", nil, UIParent)
 hidingBar.drag:SetFrameStrata("DIALOG")
-hidingBar.drag:EnableMouse(true)
 hidingBar.drag:SetSize(4, 4)
 hidingBar.drag:SetHitRectInsets(-2, -2, -2, -2)
 hidingBar.drag.bg = hidingBar.drag:CreateTexture(nil, "OVERLAY")
@@ -78,8 +77,11 @@ function hidingBar:ADDON_LOADED(addonName)
 		self.db = HidingBarDB
 		self.db.config = self.db.config or {}
 		self.config = self.db.config
-		self.config.fadeOpacity = self.config.fadeOpacity or .2
 		self.config.orientation = self.config.orientation or 0
+		self.config.fadeOpacity = self.config.fadeOpacity or .2
+		self.config.showHandler = self.config.showHandler or 0
+		self.config.showDelay = self.config.showDelay or 0
+		self.config.hideDelay = self.config.hideDelay or .75
 		self.config.size = self.config.size or 10
 		self.config.buttonSize = self.config.buttonSize or 32
 		self.config.anchor = self.config.anchor or "top"
@@ -98,6 +100,7 @@ function hidingBar:ADDON_LOADED(addonName)
 		config.hidingBar = self
 		config.config = self.config
 
+		self.drag:setShowHandler()
 		C_Timer.After(0, function() self:init() end)
 	end
 end
@@ -465,13 +468,13 @@ function hidingBar:grabMinimapAddonsButtons(t)
 			else
 				local mouseEnabled, clickable = {}
 				local function getMouseEnabled(frame)
-					for _, fchild in ipairs({frame:GetChildren()}) do
-						if fchild:IsMouseEnabled() then
-							tinsert(mouseEnabled, fchild)
-							if fchild:HasScript("OnClick") and fchild:GetScript("OnClick") then
-								clickable = true
-							end
+					if frame:IsMouseEnabled() then
+						tinsert(mouseEnabled, frame)
+						if frame:HasScript("OnClick") and frame:GetScript("OnClick") then
+							clickable = true
 						end
+					end
+					for _, fchild in ipairs({frame:GetChildren()}) do
 						getMouseEnabled(fchild)
 					end
 				end
@@ -485,11 +488,6 @@ function hidingBar:grabMinimapAddonsButtons(t)
 						frame:SetHitRectInsets(0, 0, 0, 0)
 						frame:HookScript("OnEnter", enter)
 						frame:HookScript("OnLeave", leave)
-					end
-					if child:IsMouseEnabled() then
-						self.SetHitRectInsets(child, 0, 0, 0, 0)
-						self.HookScript(child, "OnEnter", enter)
-						self.HookScript(child, "OnLeave", leave)
 					end
 
 					self.SetClipsChildren(child, true)
@@ -779,7 +777,7 @@ end
 
 
 hidingBar.drag:SetScript("OnMouseDown", function(_, button)
-	if button == "LeftButton" and not hidingBar.config.lock then
+	if button == "LeftButton" and not hidingBar.config.lock and hidingBar:IsShown() then
 		hidingBar.isDrag = true
 		hidingBar.cover:Show()
 		hidingBar:SetScript("OnUpdate", hidingBar.dragBar)
@@ -817,7 +815,6 @@ function hidingBar:enter()
 	end
 end
 hidingBar:SetScript("OnEnter", hidingBar.enter)
-hidingBar.drag:SetScript("OnEnter", enter)
 
 
 function hidingBar:hideBar(elapsed)
@@ -836,9 +833,58 @@ end
 function hidingBar:leave()
 	self.isMouse = false
 	if not self.isDrag and self:IsShown() then
-		self.timer = .75
+		self.timer = self.config.hideDelay
 		self:SetScript("OnUpdate", self.hideBar)
 	end
 end
 hidingBar:SetScript("OnLeave", hidingBar.leave)
-hidingBar.drag:SetScript("OnLeave", leave)
+
+
+function hidingBar.drag:hoverHandler()
+	if hidingBar.config.fade then
+		UIFrameFadeOut(self, hidingBar.config.showDelay, self:GetAlpha(), 1)
+	end
+end
+
+
+function hidingBar.drag:showBarDelay(elapsed)
+	self.timer = self.timer - elapsed
+	if self.timer <= 0 then
+		self:SetScript("OnUpdate", nil)
+		hidingBar:enter()
+	end
+end
+
+
+function hidingBar.drag:showOnHoverWithDelay()
+	if hidingBar:IsShown() or hidingBar.config.showDelay == 0 then
+		hidingBar:enter()
+	else
+		self:hoverHandler()
+		self.timer = hidingBar.config.showDelay
+		self:SetScript("OnUpdate", self.showBarDelay)
+	end
+end
+
+
+function hidingBar.drag:setShowHandler()
+	if hidingBar.config.showHandler == 2 then
+		self:SetScript("OnEnter", self.showOnHoverWithDelay)
+		self:SetScript("OnClick", enter)
+	elseif hidingBar.config.showHandler == 1 then
+		self:SetScript("OnEnter", self.hoverHandler)
+		self:SetScript("OnClick", enter)
+	else
+		self:SetScript("OnEnter", self.showOnHoverWithDelay)
+		self:SetScript("OnClick", nil)
+	end
+end
+
+
+hidingBar.drag:SetScript("OnLeave", function(self)
+	self:SetScript("OnUpdate", nil)
+	if hidingBar.config.fade and not hidingBar:IsShown() then
+		UIFrameFadeOut(self, hidingBar.config.showDelay, self:GetAlpha(), hidingBar.config.fadeOpacity)
+	end
+	hidingBar:leave()
+end)