+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-* Added support for Calendar, Tracker and Garrison buttons.
-* Added option to grab after n seconds
-* Added color settings.
+* Added "Show on" option.
+* Added "Delay to show" option.
+* Added "Delay to hide" option.